chore: part 4 - active caip network handling
