+++ packages/adapters/ethers/src/client.ts
@@ -892,7 +892,7 @@ export class EthersAdapter {
       }
     } else {
       this.appKit?.resetWcConnection()
-      this.appKit?.resetNetwork()
+      this.appKit?.resetNetwork(this.chainNamespace)
       this.appKit?.setAllAccounts([], this.chainNamespace)
     }
   }
