+++ packages/adapters/ethers5/src/client.ts
@@ -892,7 +892,7 @@ export class Ethers5Adapter {
       }
     } else {
       this.appKit?.resetWcConnection()
-      this.appKit?.resetNetwork()
+      this.appKit?.resetNetwork(this.chainNamespace)
       this.appKit?.setAllAccounts([], this.chainNamespace)
     }
   }
